[PRODEV-1896] GitHub action for argo deployment (#44)
Motivation
---
Need a way to apply a unique version and deploy via ArgoCD.

Modifications
---
I added an action template to apply the github.sha of the triggering
workflow as the deployment version. Then it initiates a deployment sync
for the calling app.

Results
---
MVP of ArgoCD app deployment (sync) action.