(PRODEV-5010) - Add Image Scanning to CI (#81)
NOTE: Merge this after all the permission updates have gone through.

Motivation
---
I want our images to be scanned for critical CVEs during CI, and for CI
to fail when they are detected, so we can enforce resolution of these
vulnerabilities as a part of development rather than in response to
discovery. I also want a report of the scans to be posted for reference
during resolution, and for auditing.

Modifications
---
I added image scanning, report-generation, and CI failure to the
`build-docker-images-ecr` action.

Results
---
Images are now required to be free from critical CVEs before a PR can be
merged.