Disable image scanning for the time being (#87)
Motivation
---
Image scanning is consuming WAY too much Github Runner memory to be
effective as a long term solution. It is blocking CI. We will research
and implement a different tool soon, but need to remove it for now to
unblock the pipeline.

Modifications
---
This just disables image scanning for the time being.

Results
---
CI is great again.