(PRODEV-5775) - Implement a different solution for image scanning in ci (#98)
Motivation
---
We've been without a ci-based scanning solution since Docker Scout
proved to be ineffective. We need one, for compliance and sanity.

Modifications
---
* I added trivy to our image-build steps, complete with a github actions
report template in markdown, so the things can be printed to the CI
Summary.
* I also added a mechanism by which we can ignore vulnerabilities, if we
want to, by PR'ing an update to the .trivyignores file in the actions
repo.

Results
---
CI will scan images after they are built and fail if critical or high
vulnerabilities are found ONLY if they have fixes out for them.
Unresolved vulnerabilities will be ignored.